+++ clang/test/SemaCXX/complex-div.cpp
@@ -1,3 +1,0 @@
-// RUN: %clang_cc1 -fsyntax-only %s
-
-auto f() { return 43273096 / 65536j; }